pyproject.toml
[project]
name = "thingsboard-scripts"
version = "0.1.0"
description = "Scripts to test TB features and export data"
readme = "README.md"
requires-python = ">=3.12"

dependencies = [
    "paho-mqtt>=2.1.0",
]

[dependency-groups]
dev = [
    "mypy>=1.18.2",
    "pre-commit>=4.4.0",
]
